tests/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vast-sdk-tests"
version = "0.0.0"
description = "Test suite and dependencies for Vast SDK"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
cryptography = ">=41.0"
pyOpenSSL = ">=23.0"
vastai = { path = "..", develop = true }

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "live: tests that hit the real Vast.ai API (require VAST_API_KEY)",
]
addopts = "-m 'not live' --tb=short -q"
